requirements.txt
azure-storage-blob==12.19.0
azure-ai-documentintelligence==1.0.0
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
streamlit==1.35.0
openpyxl==3.1.2 